feat(resolver): add namespace and channel awareness
This change enables  dependency resolution to choose a (namespace)
prexisting CSV to add as a step resource or if one does not exist,
choose the CSV belonging to the package's default channel.

This change also introduces wait for delete logic in installplan,
csv, and crd cleanup functions.

refactor(registry): export InMem.addPackageManifest method

refactor(resolver): add awareness of prexisting CRD owners in target
namespace

Sends a mapping of CRD to existing owners in namespace as parameter to
resolver. This allows the resolver to return an error if more than one
preexisting CSV is discovered for a given dependent CRD.

chore(vendor): stop tracking vendor directory