Annotate CSVs with the properties used during dependency resolution.
This ensures that property information is not lost after an operator
is installed, and that the properties of an operator are the same
before and after installation. Preserving properties also allows
installed operators to satisfy dependencies on properties that cannot
be inferred from a ClusterServiceVersion spec, and is a step toward
unifying installed and available operators from the perspective of
resolution.