Explain unique replaces chain requirement in channel sort errors.
The "multiple channel head" error that can be returned from channel
sorting has been a source of confusion for users. The error message
now explains that a unique replacement chain is required in order to
define the relative order between channel entries, and it also
provides an abbreviated list of all replacement chains identified, in
the form "head...tail" or "singleton".

Signed-off-by: Ben Luddy <bluddy@redhat.com>